connect the website with google firebase and implemented realtime connection with main page swapper and firestore database
